+++ v3.x/kern/i386/scb/speaker.c
@@ -54,7 +54,7 @@ static int filter(int x)
         return ret;
 }
 
-static void intr_handler(int t)
+static void intr_handler(long t)
 {
 	char data, temp;
 	int go = 0;
