+++ v3.x/kern/msg_clock/cmdclk.c
@@ -64,7 +64,7 @@ void CommandClock_Get(char *command)
 
 extern int cpu_used[];
 
-static void CommandClock_task(int t)
+static void CommandClock_task(long t)
 {
 	unsigned long command;
 	char R = 'R';
